+++ euler69.cpp
@@ -3,7 +3,7 @@
 int main(void);
 
 int main(void) {
-    vector<int> totients = totient_sieve(1000000 + 1);
+    vector<long> totients = totient_sieve(1000000 + 1);
     double max = 1.0;
     int maxn = 0;
     for (double term, n = 1; n <= 1000000; ++n) {
